[IDE] Track startup metadata in one place
Fixes VSTS #823325 The startup telemetry metadata code should be consolidated into one class
