+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.csproj
@@ -4238,6 +4238,7 @@
     <Compile Include="MonoDevelop.Ide.TypeSystem\HackyWorkspaceFilesCache.cs" />
     <Compile Include="MonoDevelop.Ide.Gui.Dialogs\NewFolderDialog.cs" />
     <Compile Include="MonoDevelop.Components.PropertyGrid\IPropertyGrid.cs" />
+    <Compile Include="MonoDevelop.Ide\IdeStartupTracker.cs" />
   </ItemGroup>
   <ItemGroup>
     <Data Include="options\DefaultEditingLayout.xml" />
