+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/IdeStartup.cs
@@ -68,10 +68,6 @@ namespace MonoDevelop.Ide
 		bool initialized;
 		static StartupInfo startupInfo;
 		static readonly int ipcBasePort = 40000;
-		static Stopwatch startupTimer = new Stopwatch ();
-		static Stopwatch startupSectionTimer = new Stopwatch ();
-		static Stopwatch timeToCodeTimer = new Stopwatch ();
-		static Dictionary<string, long> sectionTimings = new Dictionary<string, long> ();
 		static bool hideWelcomePage;
 
 		static TimeToCodeMetadata ttcMetadata;
@@ -92,12 +88,10 @@ namespace MonoDevelop.Ide
 
 			//ensure native libs initialized before we hit anything that p/invokes
 			Platform.Initialize ();
-			sectionTimings ["PlatformInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("PlatformInitialization");
 
 			GettextCatalog.Initialize ();
-			sectionTimings ["GettextInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("GettextInitialization");
 
 			LoggingService.LogInfo ("Operating System: {0}", SystemInformation.GetOperatingSystemDescription ());
 
@@ -142,8 +136,7 @@ namespace MonoDevelop.Ide
 			var args = options.RemainingArgs.ToArray ();
 			IdeTheme.InitializeGtk (BrandingService.ApplicationName, ref args);
 
-			sectionTimings["GtkInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("GtkInitialization");
 			LoggingService.LogInfo ("Using GTK+ {0}", IdeVersionInfo.GetGtkVersion ());
 
 			// XWT initialization
@@ -154,8 +147,7 @@ namespace MonoDevelop.Ide
 			Xwt.Toolkit.CurrentEngine.RegisterBackend<IExtendedTitleBarDialogBackend,GtkExtendedTitleBarDialogBackend> ();
 			IdeTheme.SetupXwtTheme ();
 
-			sectionTimings["XwtInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("XwtInitialization");
 
 			//default to Windows IME on Windows
 			if (Platform.IsWindows && GtkWorkarounds.GtkMinorVersion >= 16) {
@@ -174,14 +166,12 @@ namespace MonoDevelop.Ide
 			SynchronizationContext.SetSynchronizationContext (DispatchService.SynchronizationContext);
 			Runtime.MainSynchronizationContext = SynchronizationContext.Current;
 
-			sectionTimings["DispatchInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("DispatchInitialization");
 
 			// Initialize Roslyn's synchronization context
 			RoslynServices.RoslynService.Initialize ();
 
-			sectionTimings["RoslynInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("RoslynInitialization");
 
 			AddinManager.AddinLoadError += OnAddinError;
 
@@ -203,8 +193,7 @@ namespace MonoDevelop.Ide
 			Counters.Initialization.Trace ("Initializing Runtime");
 			Runtime.Initialize (true);
 
-			sectionTimings ["RuntimeInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("RuntimeInitialization");
 
 			bool restartRequested = PropertyService.Get ("MonoDevelop.Core.RestartRequested", false);
 			startupInfo.Restarted = restartRequested;
@@ -216,8 +205,7 @@ namespace MonoDevelop.Ide
 
 			IdeTheme.SetupGtkTheme ();
 
-			sectionTimings["ThemeInitialized"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("ThemeInitialized");
 
 			ProgressMonitor monitor = new MonoDevelop.Core.ProgressMonitoring.ConsoleProgressMonitor ();
 			
@@ -227,8 +215,7 @@ namespace MonoDevelop.Ide
 			Counters.Initialization.Trace ("Initializing Platform Service");
 			DesktopService.Initialize ();
 
-			sectionTimings["PlatformInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("PlatformInitialization");
 
 			monitor.Step (1);
 
@@ -262,8 +249,7 @@ namespace MonoDevelop.Ide
 
 			CheckFileWatcher ();
 
-			sectionTimings["FileWatcherInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("FileWatcherInitialization");
 
 			Exception error = null;
 			int reportedFailures = 0;
@@ -273,8 +259,7 @@ namespace MonoDevelop.Ide
 				//force initialisation before the workbench so that it can register stock icons for GTK before they get requested
 				ImageService.Initialize ();
 
-				sectionTimings ["ImageInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-				startupSectionTimer.Restart ();
+				IdeStartupTracker.StartupTracker.MarkSection ("ImageInitialization");
 
 				// If we display an error dialog before the main workbench window on OS X then a second application menu is created
 				// which is then replaced with a second empty Apple menu.
@@ -283,8 +268,7 @@ namespace MonoDevelop.Ide
 
 				hideWelcomePage = options.NoStartWindow || startupInfo.HasFiles || IdeApp.Preferences.StartupBehaviour.Value != OnStartupBehaviour.ShowStartWindow;
 				IdeApp.Initialize (monitor, hideWelcomePage);
-				sectionTimings ["AppInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-				startupSectionTimer.Restart ();
+				IdeStartupTracker.StartupTracker.MarkSection ("AppInitialization");
 
 				if (errorsList.Count > 0) {
 					using (AddinLoadErrorDialog dlg = new AddinLoadErrorDialog (errorsList.ToArray (), false)) {
@@ -339,8 +323,7 @@ namespace MonoDevelop.Ide
 			errorsList = null;
 			AddinManager.AddinLoadError -= OnAddinError;
 
-			sectionTimings["BasicInitializationCompleted"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("BasicInitializationCompleted");
 
 			// FIXME: we should probably track the last 'selected' one
 			// and do this more cleanly
@@ -352,15 +335,13 @@ namespace MonoDevelop.Ide
 				// Socket already in use
 			}
 
-			sectionTimings["SocketInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("SocketInitialization");
 
 			initialized = true;
 			MessageService.RootWindow = IdeApp.Workbench.RootWindow;
 			Xwt.MessageDialog.RootWindow = Xwt.Toolkit.CurrentEngine.WrapWindow (IdeApp.Workbench.RootWindow);
 
-			sectionTimings["WindowOpened"] = startupSectionTimer.ElapsedMilliseconds;
-			startupSectionTimer.Restart ();
+			IdeStartupTracker.StartupTracker.MarkSection ("WindowOpened");
 
 			Thread.CurrentThread.Name = "GUI Thread";
 			Counters.Initialization.Trace ("Running IdeApp");
@@ -375,11 +356,8 @@ namespace MonoDevelop.Ide
 			// we show the main window.
 			DispatchService.RunPendingEvents ();
 
-			sectionTimings ["PumpEventLoop"] = startupSectionTimer.ElapsedMilliseconds;
-			startupTimer.Stop ();
-			startupSectionTimer.Stop ();
-
-			CreateStartupMetadata (startupInfo, sectionTimings);
+			IdeStartupTracker.StartupTracker.MarkSection ("PumpEventLoop");
+			IdeStartupTracker.StartupTracker.Stop (startupInfo);
 
 			GLib.Idle.Add (OnIdle);
 			IdeApp.Run ();
@@ -453,19 +431,6 @@ namespace MonoDevelop.Ide
 			}
 
 			return false;
-		}
-
-		void CreateStartupMetadata (StartupInfo si, Dictionary<string, long> timings)
-		{
-			var result = DesktopService.PlatformTelemetry;
-			if (result == null) {
-				return;
-			}
-
-			var startupMetadata = GetStartupMetadata (si, result, timings);
-			Counters.Startup.Inc (startupMetadata);
-
-			IdeApp.OnStartupCompleted (startupMetadata, timeToCodeTimer);
 		}
 
 		static DateTime lastIdle;
@@ -745,13 +710,8 @@ namespace MonoDevelop.Ide
 		
 		public static int Main (string[] args, IdeCustomizer customizer = null)
 		{
-			// Using a Stopwatch instead of a TimerCounter since calling
-			// TimerCounter.BeginTiming here would occur before any timer
-			// handlers can be registered. So instead the startup duration is
-			// set as a metadata property on the Counters.Startup counter.
-			startupTimer.Start ();
-			startupSectionTimer.Start ();
-			timeToCodeTimer.Start ();
+
+			IdeStartupTracker.StartupTracker.Start ();
 
 			var options = MonoDevelopOptions.Parse (args);
 			if (options.ShowHelp || options.Error != null)
@@ -776,8 +736,7 @@ namespace MonoDevelop.Ide
 					exename = exename.ToLower ();
 				Runtime.SetProcessName (exename);
 
-				sectionTimings ["mainInitialization"] = startupSectionTimer.ElapsedMilliseconds;
-				startupSectionTimer.Restart ();
+				IdeStartupTracker.StartupTracker.MarkSection ("mainInitialization");
 
 				var app = new IdeStartup ();
 				ret = app.Run (options);
@@ -819,49 +778,6 @@ namespace MonoDevelop.Ide
 				}
 			}
 			return null;
-		}
-
-		enum StartupType
-		{
-			Normal = 0x0,
-			ConfigurationChange = 0x1,
-			FirstLaunch = 0x2,
-			DebuggerPresent = 0x10,
-			CommandExecuted = 0x20,
-			LaunchedAsDebugger = 0x40,
-			FirstLaunchSetup = 0x80,
-
-			// Monodevelop specific
-			FirstLaunchAfterUpgrade = 0x10000
-		}
-
-		static StartupMetadata GetStartupMetadata (StartupInfo startupInfo, IPlatformTelemetryDetails platformDetails, Dictionary<string, long> timings)
-		{
-			var assetType = StartupAssetType.FromStartupInfo (startupInfo);
-			StartupType startupType = StartupType.Normal;
-
-			if (startupInfo.Restarted && !IdeApp.IsInitialRunAfterUpgrade) {
-				startupType = StartupType.ConfigurationChange; // Assume a restart without upgrading was the result of a config change
-			} else if (IdeApp.IsInitialRun) {
-				startupType = StartupType.FirstLaunch;
-			} else if (IdeApp.IsInitialRunAfterUpgrade) {
-				startupType = StartupType.FirstLaunchAfterUpgrade;
-			} else if (Debugger.IsAttached) {
-				startupType = StartupType.DebuggerPresent;
-			}
-
-			return new StartupMetadata {
-				CorrectedStartupTime = startupTimer.ElapsedMilliseconds,
-				StartupType = Convert.ToInt32 (startupType),
-				AssetTypeId = assetType.Id,
-				AssetTypeName = assetType.Name,
-				IsInitialRun = IdeApp.IsInitialRun,
-				IsInitialRunAfterUpgrade = IdeApp.IsInitialRunAfterUpgrade,
-				TimeSinceMachineStart = (long)platformDetails.TimeSinceMachineStart.TotalMilliseconds,
-				TimeSinceLogin = (long)platformDetails.TimeSinceLogin.TotalMilliseconds,
-				Timings = timings,
-				StartupBehaviour = IdeApp.Preferences.StartupBehaviour.Value
-			};
 		}
 
 		internal static OpenWorkspaceItemMetadata GetOpenWorkspaceOnStartupMetadata ()
